added support for ROM over 0xFF

+++ src/hex_table.rs
@@ -8,14 +8,14 @@ impl HexTable {
     pub fn new(ins:& Vec<Instruction>) -> HexTable{
         let mut max=0;
         for inst in ins {
-            if inst.offset()as i16+inst.len()>max{
-                max = inst.offset()as i16+inst.len();
+            if inst.offset()as i32+inst.len()>max{
+                max = inst.offset()as i32+inst.len();
             }
         }
         let table =  vec![0u8; max as usize];
         HexTable{table:table}
     }
-    pub fn update(&mut self, offset:u8, hex:&Vec<u8>) {
+    pub fn update(&mut self, offset:u16, hex:&Vec<u8>) {
         for i in 0..(hex.len()) {
             self.table[offset as usize+i] = hex[i];
         }
